Split more renderer command and row modules
